United two services, removed duplicate code
