+++ Api/HappyTravel.Edo.Api.xml
@@ -107,7 +107,7 @@
         </member>
         <member name="M:HappyTravel.Edo.Api.Controllers.AdministratorControllers.BookingsController.Discard(System.Int32)">
             <summary>
-                Cancel accommodation booking by admin.
+                Cancels accommodation booking by admin.
             </summary>
             <param name="bookingId">Id of booking to cancel</param>
             <returns></returns>
@@ -1706,7 +1706,7 @@
             E-mail for billing operations
             </summary>
         </member>
-        <member name="P:HappyTravel.Edo.Api.Models.Agents.CounterpartyInfo.HasContract">
+        <member name="P:HappyTravel.Edo.Api.Models.Agents.CounterpartyInfo.IsContractUploaded">
             <summary>
             True if contract is loaded to counterparty
             </summary>
@@ -2507,6 +2507,11 @@
             Gets all booking info of the current agent
             </summary>
             <returns>List of the slim booking models </returns>
+        </member>
+        <member name="T:HappyTravel.Edo.Api.Services.Accommodations.Bookings.Management.IBookingManagementService">
+            <summary>
+            Common booking management service to be used by agents, administrators, service accounts
+            </summary>
         </member>
         <member name="M:HappyTravel.Edo.Api.Services.Accommodations.Bookings.ResponseProcessing.BookingAuditLogService.Add(HappyTravel.EdoContracts.Accommodations.Booking,HappyTravel.Edo.Data.Bookings.Booking)">
             <summary>
